refactor: consolidate utilities and clarify file structure
- Move src/utils/file-utils.ts to src/core/utils/file-utils.ts
- Rename src/mcp/services.ts to src/mcp/service-init.ts for clarity
- Update all import paths accordingly

Improves codebase organization:
- All core utilities now under src/core/utils/
- service-init.ts name clarifies its purpose vs services/ folder
