+++ src/mcp/service-init.ts
